Rename `ConfigurableDataTable` to `EntityDataTable`.
