+++ graylog2-web-interface/src/components/common/EntityDataTable/TableHead.tsx
