+++ graylog2-web-interface/src/components/common/EntityDataTable/types.ts
