+++ graylog2-web-interface/src/components/common/index.tsx
@@ -29,7 +29,7 @@ export { default as ColorPicker } from './ColorPicker';
 export { default as ColorPickerPopover } from './ColorPickerPopover';
 export { default as ConfirmDialog } from './ConfirmDialog';
 export { default as ConfirmLeaveDialog } from './ConfirmLeaveDialog';
-export { default as ConfigurableDataTable } from './ConfigurableDataTable';
+export { default as EntityDataTable } from './EntityDataTable';
 export { default as ContentHeadRow } from './ContentHeadRow';
 export { default as ControlledTableList } from './ControlledTableList';
 export { default as CopyToClipboardCapture } from './CopyToClipboardCapture';
